requirements.txt
streamlit==1.38.0
duckduckgo-search==5.3.1
openai==1.40.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
pydantic==2.8.0
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
pytest==8.0.0
